tests/selenium/requirements.txt
pytest==8.4.1
selenium==4.28.1
webdriver-manager==4.0.2
requests==2.32.3
python-dotenv==1.0.1
packaging<26
virtualenv==20.31.2
